fix(audit low): i32 week overflow, extract_bits UB shift, navbit CRC empty-input panic
Three latent correctness/robustness traps from the audit's Low section, each now
covered by a unit test:

1. Week*604800 was computed in i32 (gnsstime.rs gps_time_to_utc, satellite_param.rs
   GLONASS time, ×2) and only then cast to u32. i32::MAX is 2.147e9, so weeks > 3550
   overflow — debug panic / release wraparound — corrupting the leap-second lookup
   near GPS week rollover (~2048). Compute in i64 first. Identical result for the
   current week 2369; test pins week 5000 (would panic before).

2. extract_bits used `(1u32 << length) - 1`, undefined for length >= 32 (`1u32 << 32`
   panics in debug, wraps in release). The public bit extractor is re-exported and
   used by tests; current callers stay <= 24 but the boundary was a UB trap. Added a
   width-safe mask mirroring sign_extend. Test covers length == 32.

3. NavBit::crc24q_encode indexed bit_stream[0] with no empty guard — a latent panic,
   unlike the canonical crate::crc24q::crc24q_encode which returns 0. Added the guard.

None of these change behavior for current inputs; they remove future/boundary traps.
56 lib tests + integration suite pass, clippy clean.

+++ src/satellite_param.rs
@@ -245,8 +245,8 @@ pub fn get_satellite_param(
             adjusted_time.MilliSeconds -= 14000;
         }
         GnssSystem::GlonassSystem => {
-            // GLONASS time handling
-            let seconds = (time.Week * 604800 + time.MilliSeconds / 1000) as u32;
+            // GLONASS time handling. Compute in i64: Week*604800 overflows i32 for weeks > 3550.
+            let seconds = ((time.Week as i64) * 604800 + (time.MilliSeconds as i64) / 1000) as u32;
             let leap_second = get_leap_second(seconds);
             adjusted_time.MilliSeconds -= leap_second * 1000;
         }
@@ -1243,8 +1243,8 @@ pub fn get_satellite_param_with_prediction(
             adjusted_time.MilliSeconds -= 14000;
         }
         GnssSystem::GlonassSystem => {
-            // GLONASS time handling
-            let seconds = (time.Week * 604800 + time.MilliSeconds / 1000) as u32;
+            // GLONASS time handling. Compute in i64: Week*604800 overflows i32 for weeks > 3550.
+            let seconds = ((time.Week as i64) * 604800 + (time.MilliSeconds as i64) / 1000) as u32;
             let leap_second = get_leap_second(seconds);
             adjusted_time.MilliSeconds -= leap_second * 1000;
         }
